Extract endpoint metadata YAML serialization
