+++ compiler/ballerina-lang/src/main/java/io/ballerina/projects/JBallerinaBackend.java
@@ -17,7 +17,6 @@
  */
 package io.ballerina.projects;
 
-import com.google.gson.Gson;
 import io.ballerina.projects.environment.PackageCache;
 import io.ballerina.projects.environment.ProjectEnvironment;
 import io.ballerina.projects.internal.DefaultDiagnosticResult;
@@ -25,6 +24,7 @@ import io.ballerina.projects.internal.PackageDiagnostic;
 import io.ballerina.projects.internal.ProjectDiagnosticErrorCode;
 import io.ballerina.projects.internal.model.Target;
 import io.ballerina.projects.plugins.EndpointMetaInfo;
+import io.ballerina.projects.util.EndpointMetadataYamlUtil;
 import io.ballerina.projects.util.ProjectConstants;
 import io.ballerina.projects.util.ProjectUtils;
 import io.ballerina.tools.diagnostics.Diagnostic;
@@ -93,16 +93,9 @@ public class JBallerinaBackend extends CompilerBackend {
 
     private static final String ARTIFACT_DIR = "artifact";
     private static final String ENDPOINTS_FILE = "endpoints.yaml";
-    private static final String ENDPOINTS_KEY = "endpoints";
-    private static final String NAME_KEY = "name";
-    private static final String PORT_KEY = "port";
-    private static final String BASE_PATH_KEY = "basePath";
-    private static final String TYPE_KEY = "type";
-    private static final String SCHEMA_PATH_KEY = "schemaPath";
     private static final String JAR_FILE_EXTENSION = ".jar";
     private static final String TEST_JAR_FILE_NAME_SUFFIX = "-testable";
     private static final String JAR_FILE_NAME_SUFFIX = "";
-    private static final Gson GSON = new Gson();
     private static final HashSet<String> excludeExtensions = new HashSet<>(Lists.of("DSA", "SF"));
     private static final String OS = System.getProperty("os.name").toLowerCase(Locale.getDefault());
     public static final String JAR_NAME_SEPARATOR = "-";
@@ -295,41 +288,11 @@ public class JBallerinaBackend extends CompilerBackend {
         Path artifactDir = packageContext.project().targetDir().resolve(ARTIFACT_DIR);
         try {
             Files.createDirectories(artifactDir);
-            Files.writeString(artifactDir.resolve(ENDPOINTS_FILE), toYaml(endpointMetadata), StandardCharsets.UTF_8);
+            Files.writeString(artifactDir.resolve(ENDPOINTS_FILE), EndpointMetadataYamlUtil.toYaml(endpointMetadata),
+                    StandardCharsets.UTF_8);
         } catch (IOException e) {
             throw new ProjectException("unable to export endpoint metadata: " + e.getMessage(), e);
         }
-    }
-
-    private static String toYaml(List<EndpointMetaInfo> endpoints) {
-        StringBuilder content = new StringBuilder();
-        content.append(ENDPOINTS_KEY).append(':').append('\n');
-        for (EndpointMetaInfo endpoint : sorted(endpoints)) {
-            content.append("  - ").append(NAME_KEY).append(": ").append(quote(endpoint.name()))
-                    .append('\n');
-            content.append("    ").append(PORT_KEY).append(": ").append(endpoint.port()).append('\n');
-            content.append("    ").append(BASE_PATH_KEY).append(": ").append(quote(endpoint.basePath()))
-                    .append('\n');
-            content.append("    ").append(TYPE_KEY).append(": ").append(quote(endpoint.type()))
-                    .append('\n');
-            content.append("    ").append(SCHEMA_PATH_KEY).append(": ").append(quote(endpoint.schemaPath()))
-                    .append('\n');
-        }
-        return content.toString();
-    }
-
-    private static List<EndpointMetaInfo> sorted(List<EndpointMetaInfo> endpoints) {
-        return endpoints.stream()
-                .sorted(Comparator.comparing(EndpointMetaInfo::name)
-                        .thenComparing(EndpointMetaInfo::type)
-                        .thenComparing(EndpointMetaInfo::basePath)
-                        .thenComparingInt(EndpointMetaInfo::port)
-                        .thenComparing(EndpointMetaInfo::schemaPath))
-                .toList();
-    }
-
-    private static String quote(String value) {
-        return GSON.toJson(value);
     }
 
     public List<Diagnostic> notifyCompilationCompletion(Path filePath, BalCommand balCommand) {
