+++ compiler/ballerina-lang/src/main/java/io/ballerina/projects/util/EndpointMetadataYamlUtil.java
@@ -1,0 +1,74 @@
+/*
+ *  Copyright (c) 2026, WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
+ *
+ *  WSO2 LLC. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+package io.ballerina.projects.util;
+
+import com.google.gson.Gson;
+import io.ballerina.projects.plugins.EndpointMetaInfo;
+
+import java.util.Comparator;
+import java.util.List;
+
+/**
+ * Serializes endpoint metadata as YAML.
+ *
+ * @since 2201.13.6
+ */
+public final class EndpointMetadataYamlUtil {
+
+    private static final String ENDPOINTS_KEY = "endpoints";
+    private static final String NAME_KEY = "name";
+    private static final String PORT_KEY = "port";
+    private static final String BASE_PATH_KEY = "basePath";
+    private static final String TYPE_KEY = "type";
+    private static final String SCHEMA_PATH_KEY = "schemaPath";
+    private static final Gson GSON = new Gson();
+
+    private EndpointMetadataYamlUtil() {
+    }
+
+    public static String toYaml(List<EndpointMetaInfo> endpoints) {
+        StringBuilder content = new StringBuilder();
+        content.append(ENDPOINTS_KEY).append(':').append('\n');
+        for (EndpointMetaInfo endpoint : sorted(endpoints)) {
+            content.append("  - ").append(NAME_KEY).append(": ").append(quote(endpoint.name()))
+                    .append('\n');
+            content.append("    ").append(PORT_KEY).append(": ").append(endpoint.port()).append('\n');
+            content.append("    ").append(BASE_PATH_KEY).append(": ").append(quote(endpoint.basePath()))
+                    .append('\n');
+            content.append("    ").append(TYPE_KEY).append(": ").append(quote(endpoint.type()))
+                    .append('\n');
+            content.append("    ").append(SCHEMA_PATH_KEY).append(": ").append(quote(endpoint.schemaPath()))
+                    .append('\n');
+        }
+        return content.toString();
+    }
+
+    private static List<EndpointMetaInfo> sorted(List<EndpointMetaInfo> endpoints) {
+        return endpoints.stream()
+                .sorted(Comparator.comparing(EndpointMetaInfo::name)
+                        .thenComparing(EndpointMetaInfo::type)
+                        .thenComparing(EndpointMetaInfo::basePath)
+                        .thenComparingInt(EndpointMetaInfo::port)
+                        .thenComparing(EndpointMetaInfo::schemaPath))
+                .toList();
+    }
+
+    private static String quote(String value) {
+        return GSON.toJson(value);
+    }
+}
